Adding in a self automated function
The new function "readin" is going to read in all the codes from the document "Database.txt", tell if its a "int" number storing that in an array, then taking the user name, put that into the array of strings and pass that throughout the program. If someone is a returning user, their data should be known and be stored in the "Database.txt".